requirements.txt
discord.py
aiosqlite
matplotlib
fpdf